+++ multisheet-excel-export/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-
 ## [Version 1.0.2]() - Enhancement - 2022-07
 - Added ability to rename sheets
 - updated "cache_utils.py" file to change cache output directory from user home directory to tmp/tmp_date
+- added component to output a partitioned dataset as multisheet excel file (where each tab is a different partition)
 
 ## [Version 1.0.1](https://github.com/dataiku/dss-plugin-ml-assisted-labeling/releases/tag/v1.0.1) - New hotfix release - 2022-03
 - Changed documentation URL